kepctl: Respect cobra directory layout
