+++ cmd/kepctl/cmd/promote.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package cmd
 
 import (
 	"github.com/spf13/cobra"
@@ -22,30 +22,26 @@ import (
 	"k8s.io/enhancements/pkg/kepctl"
 )
 
-func buildCreateCommand(k *kepctl.Client) *cobra.Command {
-	opts := kepctl.CreateOpts{}
+func buildPromoteCommand(k *kepctl.Client) *cobra.Command {
+	opts := kepctl.PromoteOpts{}
 	cmd := &cobra.Command{
-		Use:     "create [KEP]",
-		Short:   "Create a new KEP",
-		Long:    "Create a new KEP using the current KEP template for the given type",
-		Example: `  kepctl create sig-architecture/000-mykep`,
+		Use:     "promote [KEP]",
+		Short:   "Promote a KEP",
+		Long:    "Promote a KEP to a new stage for a target release",
+		Example: `  kepctl promote sig-architecture/000-mykep --stage beta --release v1.20`,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			return opts.Validate(args)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return k.Create(&opts)
+			return k.Promote(&opts)
 		},
 	}
 
 	f := cmd.Flags()
-	f.StringVar(&opts.Title, "title", "", "KEP Title")
-	f.StringArrayVar(&opts.Authors, "authors", []string{}, "Authors")
-	f.StringArrayVar(&opts.Reviewers, "reviewers", []string{}, "Reviewers")
-	f.StringVar(&opts.Type, "type", "feature", "KEP Type")
-	f.StringVarP(&opts.State, "state", "s", "provisional", "KEP State")
-	f.StringArrayVar(&opts.SIGS, "sigs", []string{}, "Participating SIGs")
-	f.StringArrayVar(&opts.PRRApprovers, "prr-approver", []string{}, "PRR Approver")
+	f.StringVarP(&opts.Stage, "stage", "s", "", "KEP Stage")
+	f.StringVarP(&opts.Release, "release", "r", "", "Target Release")
 
 	addRepoPathFlag(f, &opts.CommonArgs)
+
 	return cmd
 }
